split bridge logic
