+++ package.json
@@ -4,7 +4,9 @@
   "description": "A MetaMask compatible keyring, for ledger hardware wallets",
   "main": "index.js",
   "files": [
-    "index.js"
+    "index.js",
+    "ledger-keyring.js",
+    "ledger-keyring-mv2.js"
   ],
   "engines": {
     "node": ">=14.0.0"
